model/requirements.txt
azure-core==1.26.2
azure-cosmos==4.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
numpy==1.24.1
pandas==1.5.3
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.14
